+++ src/certdate.h
@@ -4,27 +4,23 @@
  * in file LICENSE that is included with this distribution.
  */
 /**
- * The certificate time functions
+ * The certificate date helper.
  *
- *   certtime.h
+ * The time_t <-> ASN1_TIME conversions live in certtime.h; this file adds the
+ * CertDate value type and the duration/calendar helpers built on top of them.
+ *
+ *   certdate.h
  *
  */
-#ifndef CERTTIME_H
-#define CERTTIME_H
+#ifndef CERTDATE_H
+#define CERTDATE_H
 
+#include "certtime.h"
 #include "ownedptr.h"
-
-#define CERT_TIME_FORMAT "%a %b %d %H:%M:%S %Y UTC"
 
 DEFINE_LOGGER(certs_time, "pvxs.certs.time");
 namespace pvxs {
 namespace certs {
-
-///////////// OCSP RESPONSE ERRORS
-class CertTimeParseException final : public std::runtime_error {
-    public:
-        explicit CertTimeParseException(const std::string& message) : std::runtime_error(message) {}
-};
 
 /**
  * @brief To create and manipulate Certificate Dates
@@ -43,13 +39,13 @@ struct CertDate {
     CertDate() = default;
 
     // Constructor from time_t
-    CertDate(const std::time_t& time) : t(time), s(toString(time)) {} // NOLINT(*-explicit-constructor)
+    CertDate(const std::time_t& time) : t(time), s(timeTToString(time)) {} // NOLINT(*-explicit-constructor)
     // Constructor from ASN1_TIME*
-    CertDate(const ASN1_TIME* time) : t(asn1TimeToTimeT(time)), s(toString(t)) {}// NOLINT(*-explicit-constructor), ReSharper disable once CppNonExplicitConvertingConstructor
+    CertDate(const ASN1_TIME* time) : t(asn1TimeToTimeT(time)), s(timeTToString(t)) {}// NOLINT(*-explicit-constructor), ReSharper disable once CppNonExplicitConvertingConstructor
     // Constructor from ossl_ptr<ASN1_TIME>
-    CertDate(const ossl_ptr<ASN1_TIME>& time) : t(asn1TimeToTimeT(time.get())), s(toString(t)) {}// NOLINT(*-explicit-constructor)
+    CertDate(const ossl_ptr<ASN1_TIME>& time) : t(asn1TimeToTimeT(time.get())), s(timeTToString(t)) {}// NOLINT(*-explicit-constructor)
     // Constructor from time string
-    CertDate(const std::string& time_string) : t(toTimeT(time_string)), s(CertDate(t).s) {}// NOLINT(*-explicit-constructor)
+    CertDate(const std::string& time_string) : t(stringToTimeT(time_string)), s(timeTToString(t)) {}// NOLINT(*-explicit-constructor)
 
     // Define the comparison operator
     bool operator==(const CertDate rhs) const { return this->t == rhs.t; }
@@ -65,11 +61,7 @@ struct CertDate {
      * @brief Create an ASN1_TIME object from this StatusDate object
      * @return and ASN1_TIME object corresponding this StatusDate object
      */
-    ossl_ptr<ASN1_TIME> toAsn1_Time() const {
-        ossl_ptr<ASN1_TIME> asn1(ASN1_TIME_new());
-        ASN1_TIME_set(asn1.get(), t);
-        return asn1;
-    }
+    ossl_ptr<ASN1_TIME> toAsn1_Time() const { return timeTToAsn1Time(t); }
 
     /**
      * @brief Create an ASN1_TIME object from a StatusDate object
@@ -82,94 +74,9 @@ struct CertDate {
      * @param time ASN1_TIME* time pointer to convert
      * @return a time_t (unix time) version
      */
-    static time_t asn1TimeToTimeT(const ASN1_TIME* time) {
-        std::tm t = {};
-        if (!time) return 0;
-
-        if (ASN1_TIME_to_tm(time, &t) != 1) throw std::runtime_error("Failed to convert ASN1_TIME to tm structure");
-
-        return tmToUnixTime(t);
-    }
+    static time_t asn1TimeToTimeT(const ASN1_TIME* time) { return certs::asn1TimeToTimeT(time); }
 
    private:
-    /**
-     * @brief To format a string representation of the given time_t
-     * @param time the time_t to format
-     * @return the string representation in local time
-     */
-    static std::string toString(const std::time_t& time) {
-        // A time_t that is out of gmtime()'s representable range (for example the
-        // std::numeric_limits<time_t>::max() sentinel used for "permanently valid")
-        // yields a null std::tm.  Treat that as a never-expiring marker rather than
-        // dereferencing null.
-        const std::tm* tm = std::gmtime(&time);
-        if (!tm) {
-            return "PERMANENTLY VALID";
-        }
-        char buffer[100];
-        if (std::strftime(buffer, sizeof(buffer), CERT_TIME_FORMAT, tm)) {
-            return std::string(buffer);
-        }
-        throw CertTimeParseException("Failed to format status date");
-    }
-
-    /**
-     * @brief Convert the given string to a time_t value.
-     *
-     * The string is assumed to represent a time in the UTC timezone.  The
-     * format of the string is defined by `CERT_TIME_FORMAT`.  The string is parsed,
-     * and the time_t extracted and returned.
-     *
-     * Any errors in format are signalled by raising OCSPParseExceptions as this function
-     * is called from OCSP parsing
-     *
-     * @param time_string
-     * @return
-     */
-    static time_t toTimeT(const std::string& time_string) {
-        // Read the string and parse it into std::tm
-        if (time_string.empty()) return 0;
-        std::tm tm = {};
-        std::istringstream ss(time_string);
-        ss >> std::get_time(&tm, CERT_TIME_FORMAT);
-
-        // Check if parsing was successful
-        if (ss.fail()) {
-            throw CertTimeParseException("Failed to parse date-time string.");
-        }
-
-        // Convert std::tm to time_t
-        return tmToUnixTime(tm);
-    }
-
-    /**
-     * @brief To get the time_t (unix time) from a std::tm structure
-     * @param tm std::tm structure to convert
-     * @return a time_t (unix time) version
-     */
-    static time_t tmToUnixTime(const std::tm& tm) {
-        // For accurate time calculation the start day in a year of each month
-        static const int kMonthStartDays[] = {0, 31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334};
-        const int year = 1900 + tm.tm_year;
-
-        // Calculate days up to the start of the current year
-        time_t days = (year - 1970) * 365 + (year - 1969) / 4  // Leap years
-                      - (year - 1901) / 100                    // Excluding non-leap centuries
-                      + (year - 1601) / 400;                   // Including leap centuries
-
-        // Calculate days up to the start of the current month within the current year
-        days += kMonthStartDays[tm.tm_mon];
-        if (tm.tm_mon > 1 && (year % 4 == 0 && (year % 100 != 0 || year % 400 == 0))) {
-            days += 1;  // Add one day for leap years after February
-        }
-
-        // Adjust with the current day in the month (`tm_mday` starts from 1)
-        days += tm.tm_mday - 1;
-
-        // Incorporate hours, minutes, and seconds
-        return ((days * 24 + tm.tm_hour) * 60 + tm.tm_min) * 60 + tm.tm_sec;
-    }
-
     /**
      * @brief Helper to check if we've fully reached a specific moment in time
      *
@@ -503,4 +410,4 @@ struct CertDate {
 }  // namespace certs
 }  // namespace pvxs
 
-#endif //CERTTIME_H
+#endif //CERTDATE_H
